backend/requirements.txt
amqp==2.5.2
asgiref==3.2.7
beautifulsoup4==4.9.0
billiard==3.6.3.0
bs4==0.0.1
cachetools==4.1.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.2.1
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.11.0
elastic-apm==5.6.0
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-secret-manager==0.2.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.28.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
